+++ archguard-gradle-plugin/src/main/kotlin/com/archguard/gradle/HtmlReportConfig.kt
@@ -1,0 +1,6 @@
+package com.archguard.gradle
+
+open class HtmlReportConfig {
+    var enabled: Boolean = false
+    var outputPath: String = "archguard-report.html"
+}