+++ archguard-gradle-plugin/src/main/kotlin/com/archguard/gradle/LayerConfig.kt
@@ -1,0 +1,15 @@
+package com.archguard.gradle
+
+import com.archguard.core.architecture.LayerConfig as CoreLayerConfig
+import javax.inject.Inject
+
+open class LayerConfig @Inject constructor(
+    val name: String,
+) {
+    var required: Boolean = false
+
+    fun toCoreConfig(): CoreLayerConfig = CoreLayerConfig(
+        name = name,
+        required = required,
+    )
+}